All phone flows are mostly working; need to refactor the state validation; it was not working and has been mostly disabled for now.
I have seen cases of the wrong key being interpreted EVEN if I send a hard coded key and not a soft key.  I need to investigate that.

I have found that sometimes the phone is left on a configuration screen, but that is not affecting the call placement.  I can always make sure I am back to the home screen if I send some html and then hit the setup button twice